Add full test for `BinTree` and `BinTreeNode` (#53)
* test(tree): create empty test case for `BinTree`

And modify explanation of all test cases of `BinTreeNode`

* test(tree): add test case for construct, insert of `BinTreeNode`

Modify the explanations in test cases of traverse

* fix(tree): correct errors in insert node for `BinTreeNode`

When `m_left` is not null, directly create a new object and assign it will result losing the original object and its memory. So can modify the `m_data` of old object by assigning a new value.